adapted LavadoraBotellas to build from maven
