+++ .gitignore
@@ -10,3 +10,11 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
+# IDEA project files
+.idea/
+*.iml
+
+# Eclipse project files
+.classpath
+.project
