+++ LavadoraBotellas/src/resources/imagenes/cerveza.png
