+++ LavadoraBotellas/src/resources/imagenes/fondo.png
