+++ LavadoraBotellas/src/resources/imagenes/gaseosa.png
