--- LavadoraBotellas/src/topic/topics.json
@@ -1,33 +1,0 @@
-[
-  {
-    "name" : "recepcion_botella",
-    "permission": ["recibirBotella"],
-    "fireCallback": ["finRecepcionBot"]
-  },
-  {
-    "name" : "clasificacion_botella",
-    "permission": ["clasificarBotella"],
-    "fireCallback": ["clasificadaGaseosa", "clasificadaCerveza", "clasificadaOtra"],
-    "setGuardCallback" : [["guardCerveza","guardGaseosa","guardOtra"]]
-  },
-  {
-    "name" : "lavado_cerveza",
-    "permission" : ["lavarCerveza","enjuagarCerveza","secarCerveza"],
-    "fireCallback": ["finSecadoCerveza"]
-  },
-  {
-    "name": "lavado_gaseosa",
-    "permission" : ["lavarGaseosa","secarGaseosa"],
-    "fireCallback" : ["finSecadoGaseosa"]
-  },
-  {
-    "name" : "expulsion_otra",
-    "permission": ["expulsarBotella"],
-    "fireCallback" : ["finExpulsion"]
-  },
-  {
-    "name" : "devolucion_botella",
-    "permission": ["devolverBotellaLimpia"],
-    "fireCallback" : ["finDevolucion"]
-  }
-]
